ADDED: map_ensembl_to_ucsc.py
Until native support for non-Ensembl GTFs has been added to SPECTre, an
accessory script to map UCSC transcript to Ensembl transcript
identifiers and add the necessary gene_biotype annotation to a GTF
derived from the UCSC knownGene database has been provided.

This is a (fairly) crude implementation and outputs the Ensembl-mapped
UCSC GTF with the necessary annotation attributes for SPECtre
execution. Basic stats are also provided for user review. The output
GTF has not been formally tested with SPECtre is to be used at user
discretion.

The necessary file inputs to generate the Ensembl-mapped UCSC GTF are:
1) a GTF derived from the UCSC knownGene database (see program notes
for details on how to generate this file),
2) an Ensembl GTF of the same genome version as the knownGene database,
and
3) a mapping file that connects UCSC to Ensembl identifiers (see
program notes for instructions on how to generate this file).